# docs/architecture/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: UnFranchise Marketing App API
  description: |
    API for the UnFranchise Marketing App - A corporate-controlled content sharing
    and engagement platform for UnFranchise Owners (UFOs).

    ## Authentication
    Most endpoints require Bearer token authentication via JWT.
    Include the access token in the Authorization header:
    ```
    Authorization: Bearer {access_token}
    ```

    ## Rate Limiting
    - Authentication endpoints: 5 requests per 15 minutes
    - Standard API endpoints: 100 requests per minute
    - Share endpoints: 500 requests per hour
    - Upload endpoints: 20 requests per hour

    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Pagination
    List endpoints support pagination via query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 20, max: 100)
    - `sort`: Sort field with optional `-` prefix for descending (e.g., `-createdAt`)

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.

  version: 1.0.0
  contact:
    name: API Support
    email: api-support@unfranchise.com
  license:
    name: Proprietary
servers:
  - url: https://api.unfranchise.com/v1
    description: Production server
  - url: https://staging-api.unfranchise.com/v1
...
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get new access token using refresh token from cookie
      operationId: refreshToken
      security: []
      parameters:
        - in: cookie
          name: refreshToken
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                description: New refresh token (rotated)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      expiresIn:
                        type: integer
                        example: 900
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /engagement/top-content:
    get:
      tags: [Engagement]
      summary: Get top performing content
      description: Get user's top performing content
      operationId: getTopContent
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
      responses:
        '200':
          description: Top content retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContentPerformance'

  # ===== NOTIFICATIONS =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # User & Auth
...
    ContentPerformance:
      type: object
      properties:
        contentId:
          type: string
          format: uuid
        title:
          type: string
        thumbnailUrl:
          type: string
        shares:
          type: integer
        clicks:
          type: integer
        views:
          type: integer
        clickRate:
          type: number
          format: float

    # Notifications
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...